requirements.txt
django-bootstrap3==8.2.2
django-compressor==2.1.1
django-compressor-autoprefixer==0.1.0
Django==1.11
pyBarcode==0.8b1
python-decouple==3.0
